# Encoding Performance Improvements