Refactored the struct encoder to reduce allocations and improve performance.

## Benchmarks

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkUnmarshal_MixedData_1-8          6674          6571          -1.54%
BenchmarkUnmarshal_MixedData_1000-8       764907        750495        -1.88%
BenchmarkUnmarshal_MixedData_100000-8     72274997      72978644      +0.97%
BenchmarkUnmarshal_String-8               1315          1382          +5.10%
BenchmarkUnmarshal_StringPtr-8            1666          1421          -14.71%
BenchmarkUnmarshal_Int64-8                1316          1314          -0.15%
BenchmarkUnmarshal_Float64-8              1354          1350          -0.30%
BenchmarkMarshal_MixedData_1-8            7854          3831          -51.22%
BenchmarkMarshal_MixedData_1000-8         6709016       3123703       -53.44%
BenchmarkMarshal_MixedData_100000-8       716520082     325317108     -54.60%
BenchmarkMarshal_String-8                 1282          1023          -20.20%
BenchmarkMarshal_StringPtr-8              1309          1035          -20.93%
BenchmarkMarshal_Int64-8                  1274          969           -23.94%
BenchmarkMarshal_Float64-8                1622          1326          -18.25%

benchmark                                 old allocs     new allocs     delta
BenchmarkUnmarshal_MixedData_1-8          53             53             +0.00%
BenchmarkUnmarshal_MixedData_1000-8       5704           5704           +0.00%
BenchmarkUnmarshal_MixedData_100000-8     570005         570005         +0.00%
BenchmarkUnmarshal_String-8               8              8              +0.00%
BenchmarkUnmarshal_StringPtr-8            9              9              +0.00%
BenchmarkUnmarshal_Int64-8                8              8              +0.00%
BenchmarkUnmarshal_Float64-8              9              9              +0.00%
BenchmarkMarshal_MixedData_1-8            79             48             -39.24%
BenchmarkMarshal_MixedData_1000-8         75009          44009          -41.33%
BenchmarkMarshal_MixedData_100000-8       7500015        4400015        -41.33%
BenchmarkMarshal_String-8                 12             9              -25.00%
BenchmarkMarshal_StringPtr-8              12             9              -25.00%
BenchmarkMarshal_Int64-8                  12             9              -25.00%
BenchmarkMarshal_Float64-8                15             12             -20.00%

benchmark                                 old bytes     new bytes     delta
BenchmarkUnmarshal_MixedData_1-8          5280          5280          +0.00%
BenchmarkUnmarshal_MixedData_1000-8       183316        183316        +0.00%
BenchmarkUnmarshal_MixedData_100000-8     17900777      17900778      +0.00%
BenchmarkUnmarshal_String-8               4336          4336          +0.00%
BenchmarkUnmarshal_StringPtr-8            4368          4368          +0.00%
BenchmarkUnmarshal_Int64-8                4336          4336          +0.00%
BenchmarkUnmarshal_Float64-8              4352          4352          +0.00%
BenchmarkMarshal_MixedData_1-8            7016          5248          -25.20%
BenchmarkMarshal_MixedData_1000-8         3068046       1300043       -57.63%
BenchmarkMarshal_MixedData_100000-8       289538328     112738582     -61.06%
BenchmarkMarshal_String-8                 4432          4344          -1.99%
BenchmarkMarshal_StringPtr-8              4432          4344          -1.99%
BenchmarkMarshal_Int64-8                  4424          4336          -1.99%
BenchmarkMarshal_Float64-8                4496          4400          -2.14%
```